% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_bmi.R
\name{grp_bmi}
\alias{grp_bmi}
\title{Recode BMI into CDC categories}
\usage{
grp_bmi(x, depth = "full", style = "name")
}
\arguments{
\item{x}{vector of BMI's to xform}

\item{depth}{full or part recodings}

\item{style}{names or values for factor levels}
}
\value{
The recoded factor variable
}
\description{
From RIDEO package
}
