% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.R
\name{Mean}
\alias{Mean}
\title{Average numbers with na.rm=T as the default.}
\usage{
Mean(x)
}
\arguments{
\item{x}{a vector of numbers to average.}
}
\value{
The mean of \code{x} with na.rm=T specified.
}
\description{
Average numbers with na.rm=T as the default.
}
