% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.R
\name{Median}
\alias{Median}
\title{Get the median of several numbers with na.rm=T as the default.}
\usage{
Median(x)

Median(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
The median of \code{x} with na.rm=T specified.

The median of \code{x} with na.rm=T specified.
}
\description{
Get the median of several numbers with na.rm=T as the default.

Get the median of several numbers with na.rm=T as the default.
}
