% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namelook.R
\name{namelook}
\alias{namelook}
\title{Lookup Name Function to Find Variable Names}
\usage{
namelook(x, dt)
}
\arguments{
\item{x}{string to look for.}

\item{dt}{dataframe to look in}
}
\value{
The names of the variables in the dataframe
}
\description{
Lookup Name Function to Find Variable Names
}
