% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seven11-package.R
\docType{package}
\name{seven11-package}
\alias{seven11}
\alias{seven11-package}
\title{seven11: A Convenience Store of Packages}
\description{
A bunch of self-written convenience functions for R.
}
\author{
\strong{Maintainer}: David Eagle \email{david.eagle@duke.edu} (\href{https://orcid.org/0000-0003-4909-9497}{ORCID})

}
\keyword{internal}
