% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.R
\name{Sum}
\alias{Sum}
\title{Sum numbers, with na.rm option changed}
\usage{
Sum(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The sum of \code{x} with na.rm=T specified.
}
\description{
Sum numbers, with na.rm option changed
}
